==> notebooks/4_Scrape_IDEO.ipynb <==
# Scrape IDEO Journal

# Jupyter magic
%load_ext autoreload
%autoreload 2
%matplotlib inline

# imports
import requests
from bs4 import BeautifulSoup
import re
import time
import random
import pandas as pd
# from seleniumrequests import Chrome
import pathlib

save_dir = "../1_data/IDEO_journal"
pathlib.Path(save_dir).exists()

## Create index of URLs for journal posts

url = "https://www.ideo.com/journal"


response = requests.get(url)
print(response)
soup = BeautifulSoup(response.text, 'html.parser')
links1 = [a.get('href') for a in soup.find_all('a')]
len(links1)

url = "https://www.ideo.com/journal/archive/2"

response = requests.get(url)
print(response)
soup = BeautifulSoup(response.text, 'html.parser')
links2 = [a.get('href') for a in soup.find_all('a')]
len(links2)

unique_links = links1 + links2
print(f'before dedupe: {len(unique_links)}')
unique_links = list(set(unique_links))
len(unique_links)

article_links = [l for l in unique_links if '/journal/' in l]
article_links.remove('/journal/archive')

full_urls = ['https://www.ideo.com' + link for link in article_links]


len(full_urls)

# Save them as backup
with open(f'{save_dir}/index.txt', 'w') as f:
    f.writelines(full_urls)

## Grab the text for each post

full_urls[0]

article_url = full_urls[0]
response = requests.get(full_url)
print(response)
soup = BeautifulSoup(response.text, 'html.parser')


full_url

title = soup.find('h1').text

junk = [
'Send this to a friend:',
 'Related Posts',
 'Send this to a friend:',
 'Say Hello.',
 'Come work with us! We are always looking for great talent to join our global teams.',
'Send this to a friend:',
]

def get_article(article_url):
    response = requests.get(article_url)
    print(response)
    soup = BeautifulSoup(response.text, 'html.parser')
    title = soup.find('h1').text
    text = [t.text for t in soup.find_all(['p', 'h2'])]
    
    for j in junk:
        text.remove(j)
    
    

    
    return title, text

title, text = get_article(full_url)


for article_url in full_urls:
    title, text = get_article(article_url)

    print(title)
    print(len(text), 'lines')
    
    fname = f"{save_dir}/{title}.txt"
    print(fname)
    with open(fname, 'w') as f:
        f.writelines(text)
