==> notebooks/7_Scrape_Seeking_Alpha.ipynb <==
# Scrape Seeking Alpha

- Seeking Alpha compiles Earnings Call transcripts
- scraping requires a seeking alpha account

# Jupyter magic
%load_ext autoreload
%autoreload 2
%matplotlib inline

# imports
import requests
from bs4 import BeautifulSoup
import re
import time
import random
import pandas as pd
from seleniumrequests import Chrome
import pathlib

save_dir = "../1_data/Seeking_Alpha"
# pathlib.Path(save_dir).mkdir()
pathlib.Path(save_dir).exists()

## Create index of URLs for earnings call transcripts

## Setup the Selenium Window for Querying

This part requries downloading a driver for selenium. I used Chrome to get this running.


!pwd

# Change this to where the chromedriver is in your environment
path_to_chromedriver = '/Users/ahale/chromedriver'

## pull the data and write a csv for an index

# list_item_df = pd.DataFrame()

# Get index of all articles
# Need to do this iteratively so that Seeking Alpha doesn't catch on

driver = Chrome(executable_path=path_to_chromedriver)
for i in range(1,2):
    url = "https://seekingalpha.com/earnings/earnings-call-transcripts/" + str(i)
    display(url)
#     driver = Chrome(executable_path=path_to_chromedriver)
    driver.get(url)
    
    page_source = driver.page_source
    soup = BeautifulSoup(page_source, 'html.parser')
    listitems= soup.findAll("li", {"class": "list-group-item article"})

    for listitem in listitems:
        link = listitem.find("a", {"class": "dashboard-article-link"})
        title = link.text
        article_url = link.get('href')

        desc = listitem.find('div')
        org = desc.find('span').find('a').get('title')
        date = re.search('•\n      (.*)\n\n        •', desc.text).group(1)
        # Add metadata to index
        list_item_df = list_item_df.append({
            'title': title
            , 'url': 'https://seekingalpha.com' + article_url
            , 'org': org
            , 'date': date
        }, ignore_index=True)
    time.sleep(random.random())
    
    # Close the chrome window so I don't end up with 1000s open on my machine
#     driver.close()
list_item_df

# list_item_df = list_item_df.drop(columns='Unnamed: 0')

# list_item_df.drop_duplicates().to_csv('../1_data/Seeking_Alpha/index.csv', index=False)

# Get the index dataframe given that it's already written
seeking_alpha_index = pd.read_csv('../1_data/Seeking_Alpha/index.csv')
seeking_alpha_index['filename'] = None
seeking_alpha_index.head()

# TODO: change index date to date
seeking_alpha_index.date.head()

### Get the text for each earnings transcript


# Need to launch the driver once so that we can
driver = Chrome(executable_path=path_to_chromedriver)

# Loop through all transcripts in the index
save_dir = '../1_data/Seeking_Alpha/'
for index, row in seeking_alpha_index.iloc[20:40].iterrows():
    
    url = row.url
    display(url)
    driver.get(url)
    page_source = driver.page_source
    soup = BeautifulSoup(page_source, 'html.parser')
    listitems= soup.findAll("p")
    
    # Get the transcript text and save it as a file
    org_name = '-'.join([i for i in re.split('[^a-zA-Z0-9\n\.]', row.org.replace('.', '')) if i])
    date = '-'.join([i for i in re.split('[^a-zA-Z0-9\n\.]', row.date.replace('.', '')) if i])
    filename = org_name + '_' + date
    seeking_alpha_index.filename[index] = f'{filename}.txt'

    transcript_text = ''
    for item in listitems:
        transcript_text = transcript_text + ' ' + item.text

    fname = f"{save_dir}/{filename}.txt"
    with open(fname, 'w') as f:
        f.writelines(transcript_text)
    time.sleep(random.random())

seeking_alpha_index.to_csv('../1_data/Seeking_Alpha/index_2.csv')

seeking_alpha_index.iloc[0:50]
